Mode nodes helpers to `Node.swift`
